requirements.txt
Django~=5.1
djangorestframework~=3.15.2
djangorestframework==3.15.0
dj-database-url==2.3.0
